requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
colorama==0.4.6
h11==0.16.0
idna==3.10
iniconfig==2.1.0
MouseInfo==0.1.3
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.4.1
pytweening==1.2.0
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
